+++ CommHandler/src/main/java/com/smartgridready/communicator/common/helper/DriverFactoryLoader.java
@@ -1,12 +1,56 @@
 package com.smartgridready.communicator.common.helper;
 
-import java.util.NoSuchElementException;
+import java.util.Iterator;
+import java.util.LinkedList;
+import java.util.List;
 import java.util.ServiceLoader;
+
+import com.smartgridready.driver.api.http.GenHttpClientFactory;
+import com.smartgridready.driver.api.messaging.GenMessagingClientFactory;
+import com.smartgridready.driver.api.messaging.model.MessagingPlatformType;
+import com.smartgridready.driver.api.modbus.GenDriverAPI4ModbusFactory;
 
 public class DriverFactoryLoader {
 
-    public static <T> T getImplementation(Class<T> clazz) throws NoSuchElementException {
-        ServiceLoader<T> services = ServiceLoader.load(clazz);
-        return services.findFirst().orElseThrow();
+    public static List<GenDriverAPI4ModbusFactory> getAllModbusDrivers() {
+        return getImplementations(GenDriverAPI4ModbusFactory.class);
     }
+
+    public static List<GenHttpClientFactory> getAllRestApiDrivers() {
+        return getImplementations(GenHttpClientFactory.class);
+    }
+
+    public static List<GenMessagingClientFactory> getAllMessagingDrivers() {
+        return getImplementations(GenMessagingClientFactory.class);
+    }
+
+    public static GenDriverAPI4ModbusFactory getModbusDriver() {
+        return getAllModbusDrivers().stream().findFirst().orElse(null);
+    }
+
+    public static GenHttpClientFactory getRestApiDriver() {
+        return getAllRestApiDrivers().stream().findFirst().orElse(null);
+    }
+
+    public static GenMessagingClientFactory getMessagingDriver(MessagingPlatformType platform) {
+        return getAllMessagingDrivers()
+            .stream()
+            .filter(f -> f.getSupportedPlatforms().contains(platform))
+            .findFirst()
+            .orElse(null);
+    }
+
+    private static <T> List<T> getImplementations(Class<T> clazz) {
+        ServiceLoader<T> services = ServiceLoader.load(clazz);
+
+        List<T> implementations = new LinkedList<>();
+        Iterator<T> iter = services.iterator();
+        while (iter.hasNext()) {
+           implementations.add(iter.next());
+        }
+
+        return implementations;
+    }
+
+    private DriverFactoryLoader() {}
 }